[Feat] sync element 분리, userTable 컴포넌트화  #22
