+++ public/administrate/main.html
@@ -23,6 +23,7 @@
 </head>
 
 <style>
+    
     .content-wrapper {
         min-width: 1400px;
         max-width: 1600px;
@@ -142,6 +143,11 @@
     #searchDropdown .dropdown-item:hover {
         background: var(--red50);
     }
+
+
+    #userInfo {
+
+    }
 </style>
 
 <body>
@@ -149,21 +155,24 @@
         <div id="sidebar"></div>
         <div class="content-container">
             <div id="profilebar"></div>
-            <div id="title"></div>
-            <div id="cardList"></div>
 
-            <div class="content-wrapper"> 
-                <div id="dashboardChart"></div>
-                <div id="userManagement"></div>
-                <div id="ptUserManagement"></div>
-                <div id="foodManagement"></div>
-                <div id="userFoodManagement"></div>
-                <div id="recommendFoodManagement"></div>
-                <div id="addFood"></div>
-                <div id="notice"></div>
-                <div id="adminManagement"></div>
-                <div id="trainerManagement"></div>
-                <div id="gymManagement"></div>
+            <div class="content-section">
+                <div id="title"></div>
+                <div id="cardList"></div>
+
+                <div class="content-wrapper"> 
+                    <div id="dashboardChart"></div>
+                    <div id="userManagement"></div>
+                    <div id="ptUserManagement"></div>
+                    <div id="foodManagement"></div>
+                    <div id="userFoodManagement"></div>
+                    <div id="recommendFoodManagement"></div>
+                    <div id="addFood"></div>
+                    <div id="notice"></div>
+                    <div id="adminManagement"></div>
+                    <div id="trainerManagement"></div>
+                    <div id="gymManagement"></div>
+                </div>
             </div>
         </div>
     </div>
